[refactor] HttpProcessor 클래스 수정
기존 MyRunnable 에서 더 의도를 알기 쉬운 클래스 명으로 수정하였습니다.
